libobs: Use size_t for obs_encoder_get_frame_size

+++ libobs/obs.h
@@ -2207,7 +2207,7 @@ EXPORT uint32_t obs_encoder_get_height(const obs_encoder_t *encoder);
 EXPORT uint32_t obs_encoder_get_sample_rate(const obs_encoder_t *encoder);
 
 /** For audio encoders, returns the frame size of the audio packet */
-EXPORT uint32_t obs_encoder_get_frame_size(const obs_encoder_t *encoder);
+EXPORT size_t obs_encoder_get_frame_size(const obs_encoder_t *encoder);
 
 /**
  * Sets the preferred video format for a video encoder.  If the encoder can use
